🚚 Rename: 이미지 저장 네이티브 폴더 분리
