+++ apps/mobile/src/native/save-image/index.ts
@@ -1,0 +1,1 @@
+export { saveImageToLibrary } from './saveImageToLibrary';
